feat(create leave request:

+++ server/controllers/users.js
@@ -1,7 +1,4 @@
-import connection from '../helpers/conn';
-
-const client = connection();
-client.connect();
+import queryDb from '../helpers/db';
 
 class UsersController {
   // @desc     Get all users
@@ -9,23 +6,12 @@ class UsersController {
   // @access   Public
   static getUsers(req, res) {
     const query = 'SELECT id, first_name, last_name, email FROM users';
-
-    UsersController.find(query, res);
-  }
-
-  static find(query, res) {
-    try {
-      client.query(query).then(results => {
-        res.status(200).json({
-          success: true,
-          count: results.rowCount,
-          message: 'retrieve all users successfully',
-          data: results.rows
-        });
-      });
-    } catch (error) {
-      res.status(400).json({ success: false, message: 'Server Error' });
-    }
+    queryDb.dbQuery(
+      res,
+      query,
+      'retrieve all users successfully',
+      'User not found'
+    );
   }
 
   // @desc     Get single user
@@ -33,29 +19,19 @@ class UsersController {
   // @access   Public
   static getUserById(req, res) {
     const id = parseInt(req.params.id);
-
-    const query = `SELECT * FROM users WHERE id = '${id}'`;
-
-    UsersController.findById(query, res);
-  }
-
-  static findById(query, res) {
-    try {
-      client.query(query).then(result => {
-        if (result.rowCount == 0) {
-          return res.status(400).json({ success: false, message: 'Not found' });
-        }
-        //res.status(200).json({ success: true, data: result.rows });
-
-        res.status(200).json({
-          success: true,
-          message: 'retrieve a single user successfully',
-          data: result.rows[0]
-        });
+    if (!Number(id)) {
+      return res.status(400).json({
+        success: false,
+        message: 'Please input a valid numeric value'
       });
-    } catch (error) {
-      res.status(400).json({ success: false, message: 'Invalid Request' });
     }
+    const query = `SELECT * FROM users WHERE id = '${id}'`;
+    queryDb.dbQuery(
+      res,
+      query,
+      'retrieve a single user successfully',
+      `Cannot find user with the id ${id}`
+    );
   }
 
   // @desc     Create new user
@@ -63,25 +39,22 @@ class UsersController {
   // @access   Private
   static createUser(req, res) {
     const { firstName, lastName, email } = req.body;
-
-    const query = {
-      text:
-        'INSERT INTO users(first_name, last_name, email) VALUES ($1, $2, $3) RETURNING *',
-      values: [firstName, lastName, email]
-    };
-    UsersController.create(req, res, query);
-  }
-
-  static create(req, res, query) {
-    try {
-      client.query(query).then(result => {
-        return res.status(201).json({
-          message: 'A new user created successfully',
-          data: result.rows[0]
-        });
-      });
-    } catch (error) {
-      return res.status(400).json({ success: false });
+    if (!firstName || !lastName || !email) {
+      res
+        .status(400)
+        .json({ success: false, message: 'Please provide complete details' });
+    } else {
+      const query = {
+        text:
+          'INSERT INTO users(first_name, last_name, email) VALUES ($1, $2, $3) RETURNING *',
+        values: [firstName, lastName, email]
+      };
+      queryDb.dbQuery(
+        res,
+        query,
+        'A new user created successfully',
+        'User not found'
+      );
     }
   }
 
@@ -90,31 +63,24 @@ class UsersController {
   // @access   Private
   static updateUser(req, res) {
     const id = parseInt(req.params.id);
-    const { firstName, lastName, email } = req.body;
-
-    const query = {
-      text:
-        'UPDATE users SET first_name = $1, last_name = $2, email = $3  WHERE id = $4 ',
-      values: [firstName, lastName, email, id]
-    };
-
-    UsersController.findByIdAndUpdate(req, res, query);
-  }
-
-  static findByIdAndUpdate(req, res, query) {
-    try {
-      client.query(query).then(result => {
-        if (result.rowCount == 0) {
-          return res.status(400).json({ success: false });
-        }
-        return res.status(200).json({
-          success: true,
-          message: 'Update Successfully',
-          data: result.rows
-        });
+    if (!Number(id)) {
+      res.status(400).json({
+        success: false,
+        message: 'Please input a valid numeric value'
       });
-    } catch (error) {
-      return res.status(400).json({ success: false });
+    } else {
+      const { firstName, lastName, email } = req.body;
+      const query = {
+        text:
+          'UPDATE users SET first_name = $1, last_name = $2, email = $3  WHERE id = $4 ',
+        values: [firstName, lastName, email, id]
+      };
+      queryDb.dbQuery(
+        res,
+        query,
+        'Update Successfully',
+        `Cannot find user with the id ${id}`
+      );
     }
   }
 
@@ -123,19 +89,19 @@ class UsersController {
   // @access   Private
   static deleteUser(req, res) {
     const id = parseInt(req.params.id);
-
-    const query = `DELETE FROM users WHERE id= '${id}'`;
-
-    UsersController.findByIdAndDelete(req, res, query);
-  }
-
-  static findByIdAndDelete(req, res, query) {
-    client.query(query).then(result => {
-      return res.status(200).json({
-        message: 'User deleted successfully',
-        data: {}
-      });
-    });
+    if (!Number(id)) {
+      res
+        .status(400)
+        .json({ success: false, message: 'Please provide a numeric valuer' });
+    } else {
+      const query = `DELETE FROM users WHERE id= '${id}'`;
+      queryDb.dbQuery(
+        res,
+        query,
+        'User deleted successfully',
+        `Cannot find user with the id ${id}`
+      );
+    }
   }
 }
